using BigDeciemal with price

+++ src/Products/ExpirableProduct.java
@@ -1,11 +1,12 @@
 package Products;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public class ExpirableProduct extends Product {
     private LocalDate expiryDate;
 
-    public ExpirableProduct(String name, double price, int quantity,LocalDate expiryDate) {
+    public ExpirableProduct(String name, BigDecimal price, int quantity, LocalDate expiryDate) {
         super(name, price, quantity);
         this.expiryDate = expiryDate;
     }
