+++ src/Products/Shippable.java
@@ -1,6 +1,8 @@
 package Products;
 
+import java.math.BigDecimal;
+
 public interface Shippable {
     String getName();
-    double getWeight();
+    BigDecimal getWeight();
 }
